+++ src/strategies/multimodal_strategy.py
@@ -265,6 +265,7 @@ class MultimodalStrategy(BaseAgentStrategy):
                     top_k=self.search_top_k,
                     max_content_chars=self.max_content_chars,
                     get_obo_token=_get_obo_token,
+                    user_context=self.user_context,
                 )
                 logging.info(
                     "[MultimodalStrategy] FoundryIQContextProvider created "